feat(nav): gear icon for settings, blocklist tab in settings, reorder tabs
- Remove Settings and Blocklist from top navbar (10 → 8 items)
- Add gear icon to header right side for Settings access
- Move Blocklist content into Settings as a new tab
- Reorder Settings tabs: General first, Logs last, logical setup flow
- /blocklist redirects to /settings (no 404 on old bookmarks)
- Default Settings tab changed from Indexers to General
- Add settings.tabs.blocklist i18n key to all locales (EN/FR/DE/NL)
