+++ web/src/App.tsx
@@ -1,4 +1,4 @@
-import { BrowserRouter, Routes, Route, NavLink, Link } from 'react-router-dom'
+import { BrowserRouter, Routes, Route, NavLink, Link, Navigate } from 'react-router-dom'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { api } from './api/client'
@@ -16,7 +16,6 @@ import SettingsPage from './pages/SettingsPage'
 import HistoryPage from './pages/HistoryPage'
 import SeriesPage from './pages/SeriesPage'
 import CalendarPage from './pages/CalendarPage'
-import BlocklistPage from './pages/BlocklistPage'
 import DiscoverPage from './pages/DiscoverPage'
 import Logo from './components/Logo'
 import { useTheme } from './theme'
@@ -30,8 +29,6 @@ const NAV_KEYS = [
   { to: '/series', key: 'series' },
   { to: '/calendar', key: 'calendar' },
   { to: '/discover', key: 'discover' },
-  { to: '/blocklist', key: 'blocklist' },
-  { to: '/settings', key: 'settings' },
 ]
 
 function Shell() {
@@ -74,6 +71,20 @@ function Shell() {
             </nav>
 
             <div className="flex items-center gap-3 flex-shrink-0">
+              <NavLink
+                to="/settings"
+                className={({ isActive }) =>
+                  `hidden lg:block p-2 rounded-md transition-colors ${
+                    isActive ? 'bg-slate-200 dark:bg-zinc-800 text-slate-900 dark:text-white' : 'text-slate-600 dark:text-zinc-400 hover:text-slate-900 dark:hover:text-white hover:bg-slate-200/50 dark:hover:bg-zinc-800/50'
+                  }`
+                }
+                title={t('nav.settings')}
+              >
+                <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24" strokeWidth={2}>
+                  <path strokeLinecap="round" strokeLinejoin="round" d="M9.594 3.94c.09-.542.56-.94 1.11-.94h2.593c.55 0 1.02.398 1.11.94l.213 1.281c.063.374.313.686.645.87.074.04.147.083.22.127.325.196.72.257 1.075.124l1.217-.456a1.125 1.125 0 0 1 1.37.49l1.296 2.247a1.125 1.125 0 0 1-.26 1.431l-1.003.827c-.293.241-.438.613-.43.992a7.723 7.723 0 0 1 0 .255c-.008.378.137.75.43.991l1.004.827c.424.35.534.955.26 1.43l-1.298 2.248a1.125 1.125 0 0 1-1.369.491l-1.217-.456c-.355-.133-.75-.072-1.076.124a6.47 6.47 0 0 1-.22.128c-.331.183-.581.495-.644.869l-.213 1.281c-.09.543-.56.941-1.11.941h-2.594c-.55 0-1.019-.398-1.11-.94l-.213-1.281c-.062-.374-.312-.686-.644-.87a6.52 6.52 0 0 1-.22-.127c-.325-.196-.72-.257-1.076-.124l-1.217.456a1.125 1.125 0 0 1-1.369-.49l-1.297-2.247a1.125 1.125 0 0 1 .26-1.431l1.004-.827c.292-.24.437-.613.43-.991a6.932 6.932 0 0 1 0-.255c.007-.38-.138-.751-.43-.992l-1.004-.827a1.125 1.125 0 0 1-.26-1.43l1.297-2.247a1.125 1.125 0 0 1 1.37-.491l1.216.456c.356.133.751.072 1.076-.124.072-.044.146-.086.22-.128.332-.183.582-.495.644-.869l.214-1.28Z" />
+                  <path strokeLinecap="round" strokeLinejoin="round" d="M15 12a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z" />
+                </svg>
+              </NavLink>
               {version && (
                 <span className="hidden lg:block text-xs text-slate-500 dark:text-zinc-600 whitespace-nowrap">
                   {/^\d+\.\d+/.test(version) ? `v${version}` : version}
@@ -121,6 +132,13 @@ function Shell() {
                   {t(`nav.${item.key}`)}
                 </NavLink>
               ))}
+              <NavLink
+                to="/settings"
+                className={mobileLinkClass}
+                onClick={() => setMenuOpen(false)}
+              >
+                {t('nav.settings')}
+              </NavLink>
             </nav>
             <div className="flex items-center justify-between px-4 py-2 border-t border-slate-200 dark:border-zinc-800">
               {version && (
@@ -153,7 +171,7 @@ function Shell() {
           <Route path="/series" element={<SeriesPage />} />
           <Route path="/calendar" element={<CalendarPage />} />
           <Route path="/discover" element={<DiscoverPage />} />
-          <Route path="/blocklist" element={<BlocklistPage />} />
+          <Route path="/blocklist" element={<Navigate to="/settings" replace />} />
           <Route path="/settings" element={<SettingsPage />} />
         </Routes>
       </main>
