+++ web/src/i18n/locales/de.json
@@ -197,7 +197,8 @@
       "import": "Import",
       "rootfolders": "Stammordner",
       "logs": "Protokolle",
-      "general": "Allgemein"
+      "general": "Allgemein",
+      "blocklist": "Blockliste"
     },
     "indexers": {
       "heading": "Indexer",
