+++ web/src/i18n/locales/nl.json
@@ -197,7 +197,8 @@
       "import": "Importeren",
       "rootfolders": "Hoofdmappen",
       "logs": "Logboeken",
-      "general": "Algemeen"
+      "general": "Algemeen",
+      "blocklist": "Blokkeerlijst"
     },
     "indexers": {
       "heading": "Indexers",
